[Refactor]ScrollView 사용 시Expo-Router 동작 호환을 위한 코드 수정 (layout -> _layout)
